Move DocumentEntry out of DocumentDatabase
