+++ Source/DafnyLanguageServer/Workspace/IDocumentEntry.cs
@@ -1,0 +1,15 @@
+using System.Threading.Tasks;
+
+namespace Microsoft.Dafny.LanguageServer.Workspace;
+
+public interface IDocumentEntry {
+  Task<DafnyDocument> ResolvedDocument { get; }
+  Task<DafnyDocument> TranslatedDocument { get; }
+  DafnyDocument LastPublishedDocument { get; }
+  Task<DafnyDocument> LastDocument { get; }
+  public bool Idle { get; }
+  void RestartVerification();
+
+  void EndVerification();
+
+}
